Show enchantments on localized card and enchantment pages
The Enchantments tab and the enchantment Cards tab only wired the manifest on the English routes, so they were missing on every /{lang}/ page. Pass the manifest on the localized routes too (names and renders are already localized via the API and enchantedCardUrl). Extracts the manifest lookups into lib/card-enchantments to share across all four pages.
