+++ frontend/app/enchantments/[id]/page.tsx
@@ -5,23 +5,7 @@ import JsonLd from "@/app/components/JsonLd";
 import { buildDetailPageJsonLd, buildFAQPageJsonLd } from "@/lib/jsonld";
 import { redirectMissingEntity } from "@/lib/redirect-helpers";
 import { imageUrl } from "@/lib/image-url";
-// Card -> valid enchantments manifest (from the render export). Imported
-// server-side and inverted here so we can show example renders per enchantment.
-import cardEnchantmentsManifest from "@/lib/card-enchantments.json";
-
-const CARD_SAMPLE = 48;
-
-function cardsForEnchantment(enchId: string): { cardIds: string[]; total: number } {
-  const ench = enchId.toLowerCase();
-  const all: string[] = [];
-  for (const [cardId, enchs] of Object.entries(
-    cardEnchantmentsManifest as Record<string, string[]>,
-  )) {
-    if (enchs.includes(ench)) all.push(cardId);
-  }
-  all.sort();
-  return { cardIds: all.slice(0, CARD_SAMPLE), total: all.length };
-}
+import { cardsForEnchantment } from "@/lib/card-enchantments";
 
 const API_INTERNAL = process.env.API_INTERNAL_URL || process.env.NEXT_PUBLIC_API_URL || "http://localhost:8000";
 const API_PUBLIC = process.env.NEXT_PUBLIC_SITE_URL || process.env.NEXT_PUBLIC_API_URL || "";
